Extract modals behavior into a mixin
